requirements.txt
beautifulsoup4==4.9.1
Bottleneck==1.3.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
fastai==1.0.61
fastprogress==0.2.3
Flask==1.1.2
Flask-WTF==0.14.3
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numexpr==2.7.1
numpy==1.18.4
nvidia-ml-py3==7.352.0
packaging==20.4
pandas==1.0.4
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
scipy==1.4.1
six==1.15.0
soupsieve==2.0.1
torch==1.5.0+cpu
torchvision==0.6.0+cpu
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
